[UNTESTEDGENERATOR] changed inheritance to composition
